Fixed script and simplified files structure
