--- bin/ColorScript/ColorScriptExecutable/.gitignore
@@ -1,4 +1,0 @@
-.DS_Store
-/.build
-/Packages
-/*.xcodeproj
