--- bin/ColorScript/ColorScriptExecutable/Package.resolved
@@ -1,16 +1,0 @@
-{
-  "object": {
-    "pins": [
-      {
-        "package": "ColorScriptCore",
-        "repositoryURL": "/Users/saturninotexeira/Documents/ios-oss/bin/ColorScript/ColorScriptCore",
-        "state": {
-          "branch": null,
-          "revision": "5b017d02561a174df0a058c6090c1dbfb62ac70b",
-          "version": "1.0.0"
-        }
-      }
-    ]
-  },
-  "version": 1
-}
