+++ bin/ColorScript/Sources/ColorScript/main.swift
@@ -1,0 +1,10 @@
+import Foundation
+import ColorScriptCore
+
+let colorScript = ColorScript()
+
+do {
+  try colorScript.run()
+} catch {
+  print("Error")
+}
