+++ bin/ColorScript/Tests/ColorScriptTests/XCTestManifests.swift
