+++ bin/ColorScript/Tests/LinuxMain.swift
